move types to models
