+++ app/src/ui/publish-repository/publish-repository.tsx
@@ -9,42 +9,10 @@ import { merge } from '../../lib/merge'
 import { caseInsensitiveCompare } from '../../lib/compare'
 import { sanitizedRepositoryName } from '../add-repository/sanitized-repository-name'
 import { Octicon, OcticonSymbol } from '../octicons'
-
-export interface IGHEPublicationSettings {
-  readonly kind: 'ghe'
-
-  /** The name to use when publishing the repository. */
-  readonly name: string
-
-  /** The repository's description. */
-  readonly description: string
-
-  /** Should the repository be private? */
-  readonly private: boolean
-}
-
-export interface IDotcomPublicationSettings {
-  readonly kind: 'dotcom'
-
-  /** The name to use when publishing the repository. */
-  readonly name: string
-
-  /** The repository's description. */
-  readonly description: string
-
-  /** Should the repository be private? */
-  readonly private: boolean
-
-  /**
-   * The org to which this repository belongs. If null, the repository should be
-   * published as a personal repository.
-   */
-  readonly org: IAPIUser | null
-}
-
-export type RepositoryPublicationSettings =
-  | IGHEPublicationSettings
-  | IDotcomPublicationSettings
+import {
+  RepositoryPublicationSettings,
+  IDotcomPublicationSettings,
+} from '../../models/publish-settings'
 
 interface IPublishRepositoryProps {
   /** The user to use for publishing. */
